+++ servant-util/CHANGES.md
@@ -4,3 +4,5 @@ Unreleased
 * Added logging.
 * Added sorting.
 * Add pagination.
+* Add swagger tags.
+* Add swagger error responses.